requirements.txt
flask
numpy
pillow
matplotlib
scikit-image
